move: all tasks logic to tasks component and rename admin-tasks to tasks;
