+++ src/app/admin/assessments/exam-answer-form/exam-answer-form.component.ts
@@ -1,7 +1,7 @@
 import {Component, OnInit, Input, OnChanges} from '@angular/core';
 import {FormGroup, FormBuilder, Validators, FormArray} from '@angular/forms';
 import {Question} from '../../../classes/question';
-import {Task} from '../../../classes/task';
+import {Task} from '../../tasks/task';
 import {QuestionService} from '../../../services/question.service';
 
 @Component({
