+++ src/app/admin/tasks/task.service.ts
@@ -1,9 +1,8 @@
 import { Injectable } from '@angular/core';
-import { Task } from "../classes/task";
-import { HttpService } from "./http.service";
-import { Subscription } from "rxjs/Rx";
-
-var _ = require('lodash');
+import { Task } from './task';
+import { HttpService } from '../../services/http.service';
+import { Subscription } from 'rxjs/Rx';
+import * as _ from 'lodash';
 
 @Injectable()
 export class TaskService {
@@ -113,7 +112,7 @@ export class TaskService {
 
         _.forEach(this.tasks,
             task => {
-                if (task.category.id === categoryId) {
+                if (task.category.getId() === categoryId) {
                     categoryTasks.push(task);
                 }
             }
