+++ src/app/admin/tasks/task.ts
@@ -1,5 +1,5 @@
 import { AssessmentType } from './assessment-type';
-import { Category } from '../admin/categories/category';
+import { Category } from '../categories/category';
 
 export class Task {
