+++ src/app/app.module.ts
@@ -12,6 +12,7 @@ import { AuthService } from './services/auth.service';
 import { AuthGuard } from './core/auth.guard';
 import { HttpService } from './services/http.service';
 import { AdminService } from './services/admin.service';
+import { AssessmentTypeService } from './admin/tasks/assessment-type.service';
 
 import { AppComponent } from './app.component';
 import { LoginComponent } from './components/login/login.component';
@@ -22,12 +23,10 @@ import { AccountActivationComponent } from './components/account-activation/acco
 import { AdminComponent } from './admin/admin.component';
 import { AdminGuard } from './core/admin.guard';
 import { AdminCategoriesComponent } from './admin/categories/categories.component';
-import { AdminTasksComponent } from './admin/admin-tasks/admin-tasks.component';
+import { TasksComponent } from './admin/tasks/tasks.component';
 import { CategoryService } from './admin/categories/category.service';
 import { CategoryFormComponent } from './admin/categories/form/form.component';
-import { TaskService } from './services/task.service';
 import { QuestionService } from './services/question.service';
-import { AssessmentTypeService } from './services/assessment-type.service';
 import { AssessmentsComponent } from './admin/assessments/assessments.component';
 import { OpenAnswerFormComponent } from './admin/assessments/open-answer-form/open-answer-form.component';
 import { ExamAnswerFormComponent } from './admin/assessments/exam-answer-form/exam-answer-form.component';
@@ -36,6 +35,7 @@ import { CreateComponent } from './admin/categories/create.component';
 import { EditComponent } from './admin/categories/edit.component';
 
 import { NotificationService } from './shared/notification.service';
+import {TaskService} from './admin/tasks/task.service';
 
 @NgModule({
   declarations: [
@@ -47,7 +47,7 @@ import { NotificationService } from './shared/notification.service';
     AccountActivationComponent,
     AdminComponent,
     AdminCategoriesComponent,
-    AdminTasksComponent,
+    TasksComponent,
     CategoryFormComponent,
     AssessmentsComponent,
     OpenAnswerFormComponent,
